Add initial board support directory with SoftDevice support
The "badge" directory contains an initial ChibiOS build that launches
the SoftDevice. Code to actually use it to manipulate the radio will
be forthcoming.

By default the NRF52 port wants to use one of the RTC timers as the system
timer. We can't do that because the SoftDevice wants to monopolize the RTC
(or at least RTC0). The port has been modified to use the ARM SysTick clock
instead.

A custom linker script is provided to link both the SoftDevice and ChibiOS
images together into a single ELF file to make for easier flashing.