Fix warnings
There are a few compiler warnings due to other discrepancies between the
NRF52 contributed port code and the ChibiOS base. These have been fixed.
There were also a few warnings due to coding inconsistencies in the QEI
driver and some unused macros in the hal_lld module.

The main.c module has been cleaned up to fix warnings too. Also, we
now call sd_ble_enable() to confirm that we can enable the BLE support.
When this is enabled, we need to give the SoftDevice more RAM (up to
0x20003200, which is about 12KB). The linker script has been updated
accordingly.