Add support for EXT device
Eventually we will need support for external interrupts from GPIO pins.
This wupport is implemented for the NRF51 port, but not for NRF52. This
commit fixes that.

The GPIO interrupt mechanism in the NRF52 is largely the same as it is
in the NRF51, except it's been extended from 4 channels to 8. This means
we can have up to 8 GPIO pins configured as interrupt sources.

A sample ISR that triggers when BTN2 is pressed has been added to main.c
for testing purposes.