Fix a bug in the EXT interrupt handling logic
The hal_ext_llc.c code inherited from the NRF51 port has the same logic
error in it that the KW01 code did. It treats the mode field as a bitmask
when it should be treading it as an integer value. The result of this
was that the sample button handler was triggering twice: even though we
specified we wanted falling edge only triggering, we were getting
toggle triggering instead (interrupt firing on both falling and rising
edges).

This has now been fixed as it was in the KW01 code.

Also, all the references to nRF51 in the code have been changed to nRF52.