Fix image entry point
When debugging with gdb, we need the entry point address to point
to the Reset_Vector address. But when using the SoftDevice object,
that needs to be the SoftDevice reset vector, not the one from ChibiOS.
Previously we were setting the entry point to 0. To make things
work right, we need to manually set the softdeviceEntry symbol
to the SoftDevice reset vector. This requires tweaking the mkelf.sh
script a little bit.