Add preliminary display support
The same East Rising display from the DC25 badge is now supported on the
NRF52 board. This requires some mods to the board:

- R27 and R28 must be populated with 0 ohm resistors or jumpers
- The UICR area must be reflashed to turn P09 and P10 back into GPIOs
  (these are used for NFC by default)

A special jumper harness is required.

The display is supposed in parallel mode because the SPI controller on
the NRF52832 has a maximum clock of 8MHz, which is even slower than the
12MHz SPI bus on last year's badge. That would reduce video frame rate
to maybe 5 fps, which is terrible. With 8-bit parallel mode, we can get
maybe 14 or 15 frames/second.