Add FatFs and SD/MMC to tbe build
The FatFs code is now compiled as part of the badge build. It doesn't work
yet. The mmc_spi_lld.c module is just a template and needs to be filled in
with the appropriate SPI driver code. I need to finish wiring up my display
first before I can work on this.