Clean up GAP code
In theory, using an advertizement of 0 should keep the device advertising
indefinitely, until told to stop. In practice it seems sometimes it
stops on its own for reasons that are not entirely clear. To avoid this,
we now secify a timeout of 30 seconds, and restart advertisement each time
the timeout is triggered, _except_ in the case where we're in the connected
state. You can't advertise and be connected at the same time. Once the
connection drops, we restart advertising.