Implement joypad code as separate module
Four buttons on the NRF52 development board are now handled as up/down/left/right
joypad buttons. Unfortunately there isn't a 5th button for enter. Eventually
this will be hooked to the app event handler. For now it just prints
messages on the console to indicate button press/release events.