Adjust display driver again
The glitches are not due to timing. They're most likely due to concurrent
accesses to the GPIO registers. The cheap solution for this is to mask
interrupts while doing a screen accesses. This is not ideal, but each
access takes relatively few instructions so we don't hog the CPU for
long.