Improve graphics drawing performance
WHoever wrote the NRF52 pal code used standard drive mode for the GPIO pins.
According to the manual, the GPIO transition time is faster in high drive mode.
By setting the screen pins to high drive, we reduce the number of cycles that
we have to pause between write pin toggles from 7 to 2. This decreases the
time needed to write data to the screen substantially.