Clean up a few things
In the SPI low level driver, use a single spiSend() instead of multiple
spiExchange()s when sending a command to the SD card.

Turn off some of the debug options in ChibiOS.

Also in main.c, there is a demo for the RNG that uses a 1024 byte buffer.
We don't need to use up that much RAM.