Fix a bug in the EXT driver
Now that we're using GPIOTE channel 7 as part of the workaround for erratum 58,
we need to make sure the EXT driver's interrupt handler doesn't trip over it
by mistake. Instead of just testing for events, also check to see if the
interrupts for the channel are unmasked too.

Failing to do this would cause the system to crash when a button was
pressed because an event would show up as being triggered on channel 7
due to the activity on the SPI channel, but there wouldn't be any ISR
connected for it, so the CPU would branch into the weeds.