Change screen control pins
During testing of video playback, it was observed that the display was
garbled when DMA occurred at the same time that data was written to the
screen. Previously the 'write' and 'command/data' select pins were on
pins 9 and 10. These pins default to being assigned to the NFC radio,
but can be programmed to be GPIOs. This works, but apparently these
pins glitch in some cases. Having the write pin glitch caused additional
pixel data to be clocked out to the screen, which is what caused the
distortion.

The write and comman/data pins have been moved to pins 11 and 12. This
fixed all the playback problems.

Also, currently screen read support is not implemented. It's not clear
if we'll need it, so for now the 'read' pin can be tied high.