Add initial video player module
This is just prototype code for now. It reads a file of video pixel data
and stuffs it to the screen. A helper thread is spawned to read the data
while the foreground thread writes the data to the display. Since the
screen and SD card are on different channels, the SD card and screen
accesses can happen in parallel.

Also, display the RSSI values when printing BLE advertisement reports,
and increase the BLE event thread stack a little.