Implement some synchronization for video
In the first badge design, the video was synched to the audio, which was
synched using a timer. We don't have audio support yet, so for now we use
one of the timers to just sync the frame display.

For some reason the timer driver wasn't allowing 24-bit or 32-bit resolution for
anything except TIMER0. It doesn't look like there really is such limitation
on the NRF52 so this has been disabled. GPT2 now runs at 16MHz with 32-bit
resolution.

To make the timer interval a whole number, the video chunk size has been changed
from 24 lines to 20. This also uses less memory.

Also, add the video module to the build.