Break out async file read support into a separate module
The possibility exists that we can use async I/O in the uGFX library to
speed up image loading. But first we need a formalized API for it, as opposed
to having it buried inside the video player modules.

Also, the video player sync delay calculation was incorrect. I don't know
what I was thinking when I did it before. It seems to be better now.