Enable I2C driver
We're going to need I2C support for both the LED controller and I2S codec,
so we may as well enable it now. The driver needed a minor tweak in order
to compile. Note that we're using legacy (non-DMA) mode, but that's ok
since speed for I2C should not be critical. (At least, not for the I2S
codec. I'm not sure how critical it will be for the LEDs.) Pins 7 and 8
have been allocated for I2C clock and data.

Also, the "printf" wrapper macro has been moved to a central header file
instead of being duplicated everywhere.