Add support for 5th button, add linker set support
Pin 09 has been assigned as the enter button for the joypad. The joypad code
has been updated to include this.

Also, initial support for the orchard linker set handling has been added so
that we can have separate command and application modules. Currently only
the command support has been integrated. The random and watchdog commands
have been broken out of main.c and into their own files. The watchdog module
may be removed at some point as it's not very useful for us right now. It's
being retained to document the watchdog usage.

The random command has been updated to use the SoftDevice API when the
SoftDevice is enabled, because it blocks direct access to the RNG registers
when it's running. The hardware RNG driver is now not included in the build
in this case, which saves a little flash space.