Fold in orchard app code
The Orchard app framework has been added to tbe build. Right now there's just
a sample app (app-test.c). This should be removed eventually.

A change to the button widget has also been made to match the DC25 code.
We want to signal a button event on a press rather than a release.

Currently the test app responds to touch screen press events. Support for
radio events is present but right now is just a stub.