Optimize the display access code a little more
We can save a few instructions on each screen access by folding the assertion
of the WR pin into previous accesses. This means we need a few less machine
instructions to write pixels out to the screen, which increases the data rate
by a small amount. Display behavior appears ot be unchanged.