Upgrade ChibiOS to version 17.6.3.
The original code was checked out from the ChibiOS git repo. It's probably
better though if we stick with stable releases. This code is basically the
exact contents of the ChibiOS 17.6.3 zip file except for the community
directory. We'll keep our own ChibiOS-contrib directory instead since
the NRF52 board support is all in there.