Add preliminary chat app
This is a port of the original chat app from the DC25 badge. Currently
there's no "shout" support, meaning that the peer doesn't get signalled
to let them know someone wants to chat with them. This will be added
later. (It's a little tricky to implement it the way it was before
due to the different radio API behavior.)

Unfortunately MAC addresses are rather wide and take up a lot of space
in the list widget, which crowds out the name field. I'm not sure how to
deal with this yet.a (We can change the font but then it becomes hard
to read.)

If nothing else, this serves as an example of how radio events work.

Also: made the bleGapAdvBlockFind() API public, and fixed a bug in
the L2CAP RX notification.