Do some cleanups
Add a private version of sscanf(), cribbed from FreeBSD. The version in
libc is a resource hog. This not only uses less stack space, it also
consumes about 2KB less code.

Moved the L2CAP packet buffer out of the radio event structure. Having
it there means that the orchard event structure is 128 bytes larger,
which means every time we service an app event we have that much less
space on the stack.

Also add a separate timeout value for connections. Before we would use
the scan timeout, which is 30 seconds. This is a bit excessive. It's now
set to 6 seconds.

Also added back the screen_alert_draw() function so that the chat app
can issue some status messages.