Add the LED sign app and the hardware scroll module
Also, fix a small bug in the launcher app. For some reason with the
new uGFX library, the label banner on the launcher menu is s few pixels
too small to display all of the manteka20 font, and the lower few pixels
get chopped off. We can fix this by specifying the label height as -1,
which forces the label widged to autocalculate the right height value.