Add the Atari 2600 emulator
Right now the emulator core is built, but the app is not. The emulator
still needs work if we're going to include it. It's quite slow, and
I don't know if I can overcome that. But I'd like the code checked in so
that I can track my changes.

To make everything compile, add the following sources to the build:

app-atari.c
atari.c
badge_disp.c
badge_keyb.c

The ROMs need to be added to the SD card for this to work.