Reduce RAM usage a little
Some of the thread stacks can be reduced to save some memory. Also, the
RNG driver and instance handle were still being pulled into the build
even though that driver is supposed to be disabled when we build with
SoftDevice support.

Also, the v2600 emulator raster code has been updated to write pixels
directly to the display driver so we don't need a line buffer in RAM
anymore. Unfortunately it's still much too slow to be usable. :(