Add an initial nRF52840 BSP
This is a copy of the nRF52832 BSP. It's mostly the same, except the
linker script has been updated to specify that there's more flash and
RAM. Since there's no screen yet, the uGFX stuff is turned off for now.
The 4 joypad buttons work. The 840 DK board has a different GPIO layout
than the 832 board and not everything has been fixed up for it yet.
This is just to get the board to boot and the radio running.